+++ src/main/java/com/anthonyhilyard/iceberg/renderer/VertexConsumerSodium.java
@@ -2,8 +2,8 @@ package com.anthonyhilyard.iceberg.renderer;
 
 import com.mojang.blaze3d.vertex.VertexConsumer;
 
-//import me.jellysquid.mods.sodium.client.render.vertex.VertexBufferWriter;
+import me.jellysquid.mods.sodium.client.render.vertex.VertexBufferWriter;
 
-public interface VertexConsumerSodium extends VertexConsumer //, VertexBufferWriter 
+public interface VertexConsumerSodium extends VertexConsumer, VertexBufferWriter 
 {
 }